+++ internal/app/app.go
@@ -18,6 +18,7 @@ type Config struct {
 	TerraformPluginDir       string
 	TerraformPluginBackupDir string
 	BaseDir                  string
+	GoPath                   string
 	ProvidersCacheDir        string
 }
 
@@ -36,6 +37,7 @@ func New() *App {
 	app := &App{
 		Config: &Config{
 			BaseDir:                  BaseDir,
+			GoPath:                   GetCurrentGoPath(),
 			TerraformPluginDir:       BaseDir + DefaultTerraformPluginDir,
 			TerraformPluginBackupDir: BaseDir + DefaultTerraformPluginBackupDir,
 			ProvidersCacheDir:        BaseDir + DefaultProvidersCacheDir,